# src/bkauth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BKAuth API 文档",
        "title": "BKAuth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/apps": {
            "post": {
                "description": "creates an app with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app create",
                "operationId": "api-app-create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "App Info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createAppSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/common.AppResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/oauth/apps": {
            "post": {
                "description": "creates an oauth app with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth app"
                ],
                "summary": "oauth app create",
                "operationId": "api-oauth-app-create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "App Info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createAppSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/common.AppResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/apps/{bk_app_code}": {
            "get": {
                "description": "gets an oauth app with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth app"
                ],
                "summary": "oauth app get",
                "operationId": "api-oauth-app-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.OAuthApp"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "updates an oauth app with base info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth app"
                ],
                "summary": "oauth app update",
                "operationId": "api-oauth-app-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "App Info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateOAuthAppSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.AppResponse": {
            "type": "object",
            "properties": {
                "bk_app_code": {
                    "type": "string"
                }
            }
        },
...
        "handler.createAppSerializer": {
            "type": "object",
            "required": [
                "bk_app_code",
                "name"
            ],
            "properties": {
                "bk_app_code": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "bk_paas"
                },
                "bk_app_secret": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "bk_paas"
                },
                "description": {
                    "type": "string",
                    "example": "Platform as A Service"
                },
                "name": {
                    "type": "string",
                    "example": "BK PaaS"
                }
            }
        },
...
        "handler.updateOAuthAppSerializer": {
            "type": "object",
            "required": [
                "redirect_urls"
            ],
            "properties": {
                "redirect_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[https://example.com/",
                        " http://example.com]"
                    ]
                }
            }
        },
...
        "types.OAuthApp": {
            "type": "object",
            "properties": {
                "bk_app_code": {
                    "type": "string"
                },
                "redirect_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}